🐛 FIX: Stylelint issues fixed
